+++ sandbox/notification-socket-example/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>6.0.3 (2023-12-22)</small>
+
+* chore(deps): fix the synk high issues (#1845) ([0a536b4](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/0a536b4)), closes [#1845](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1845) [#00](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/00)
+* chore(deps): fix the synk issues (#1853) ([be24dfb](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/be24dfb)), closes [#1853](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1853) [#1851](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1851)
+
+
+
+
+
 ## <small>6.0.2 (2023-12-07)</small>
 
 **Note:** Version bump only for package @sourceloop/notification-socket-example